refactor: organize ui modules
