+++ README.md
@@ -2,6 +2,8 @@
 
 ## 最新更新
 
+- **模組重構**：將 `MonteCarloWorker` 與 `PandasModel` 移至 `src/ui` 子模組，主視窗檔案更為精簡
+
 - **蒙地卡羅模擬功能增強** ✨
   - 新增 scipy 相依性，支援核密度估計(KDE)功能
   - 實現雙模式圖表顯示：「次數模式」和「密度模式」
